+++ src/renderers/accurateAsciiRenderer/shaders/asciiAccurate.frag
