+++ src/renderers/accurateAsciiRenderer/shaders/brightnessSplit.frag
