+++ src/renderers/brightnessAsciiRenderer/shaders/asciiBrightness.frag
