+++ src/renderers/edgeAsciiRenderer/shaders/asciiEdge.frag
