+++ src/renderers/edgeAsciiRenderer/shaders/sobel.frag
